requirements.txt
langchain==0.1.14
langchain-openai==0.0.10
langchain-community==0.0.29
crewai==0.3.8
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.1.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.0
azure-cognitiveservices-speech==1.35.0
openai==1.3.9
pyttsx3==2.90
SpeechRecognition==3.10.0
python-dateutil==2.8.2
pytz==2023.3
